+++ CHANGELOG.markdown
@@ -21,6 +21,7 @@ Version 1.2.2
 * Preloading cache moved from DOM to memory for better performance.
 * Hardware acceleration enabled in Firefox to help remedy occasional flickering therein (GH-196).
 * Error message into the console in case target image properties were deemed insufficient.
+* New `rowlock` and `framelock` options allow mouse interaction lock of axis in multirow setups (GH-191).
 
 
 Version 1.2.1